message catalog descriptor not necessary to be int
git-svn-id: https://stlport.svn.sourceforge.net/svnroot/stlport/branches/STLPORT_5_0/STLport@813 01b27279-081b-0410-8cf9-d9d9e0fb0389

+++ stlport/stl/_messages_facets.h
@@ -42,7 +42,7 @@ _STLP_BEGIN_NAMESPACE
 
 class messages_base {
 public:
-  typedef int catalog;
+  typedef long catalog;
 };
 
 template <class _CharT> class messages {};
